seperate distance and correlation based mutation disturbance classes
